Move log serializer task to separate file
